Update build system
* Use recursive make to factor out haskell-specific build concerns
* Avoid use of stack and custom scripts for build docker images
* Instead, build using regular Dockerfile

Rationale:
Stack tries to automate the creation of a deployable image, however, the setup
to make this work reliably turns out to be more work, harder to investigate and
less customizable than just using a Dockerfile + Makefiles to coordinate the
build. This way we can also take advantage of make's build tree pruning.